Snake | Beta Phase (0.0.5)
Added what i mainly wanted for the game, including: Microtransactions, point based system, stopwatch, and abilities in the microtransactions. | Beta phase (0.1.0) will appear when i get the application for the code to run and when i get the rest of the stuff i need for the code done. | Things for beta phase 0.1.0: A title screen with options, quit, and modes, an options toggle for wasd and arrow key swap for whatever is better for the player, enemy snakes, multiplayer??????